Move queries logic into platform-core
