+++ plugins/platform-core/src/__test__/search.test.ts
