+++ plugins/platform-core/src/cache.ts
