+++ plugins/platform-core/src/index.ts
@@ -13,11 +13,36 @@
 // limitations under the License.
 //
 
-import { ClientService } from '@anticrm/client'
-import { Doc } from '@anticrm/core'
+import { Class, CoreProtocol, Doc, DocumentProtocol, DocumentQuery, FindOptions, Model, Ref } from '@anticrm/core'
+import { OperationProtocol } from '@anticrm/domains'
 import { Metadata, Plugin, plugin } from '@anticrm/platform'
 
-export interface CoreService extends ClientService {
+// Queries
+export type Subscriber<T> = (value: T[]) => void
+export type Unsubscribe = () => void
+
+export interface QueryResult<T extends Doc> {
+  subscribe: (run: Subscriber<T>) => Unsubscribe
+}
+/**
+ * Define operations with live queries.
+ */
+export interface QueryProtocol {
+  /**
+   * Perform query construction, it will be possible to subscribe to query results.
+   * @param _class - object class
+   * @param query - query
+   */
+  query: <T extends Doc>(_class: Ref<Class<T>>, query: DocumentQuery<T>, options?: FindOptions<T>) => QueryResult<T>
+}
+
+export type QueryUpdater<T extends Doc> = (_class: Ref<Class<T>>, query: DocumentQuery<T>, options?: FindOptions<T>) => void
+
+export interface CoreService extends CoreProtocol, DocumentProtocol, QueryProtocol, OperationProtocol {
+  getModel: () => Model
+
+  generateId: () => Ref<Doc>
+
   getUserId: () => string
 }
