+++ plugins/presentation/src/utils.ts
@@ -14,15 +14,14 @@
 //
 
 import { Class, CORE_CLASS_TYPE, Doc, DocumentQuery, FindOptions, Mixin, Obj, Ref, StringProperty } from '@anticrm/core'
-import { getContext, onDestroy } from 'svelte'
-import core, { CoreService } from '@anticrm/platform-core'
-import { VDoc, Unsubscribe } from '@anticrm/domains'
-import { AnyComponent, CONTEXT_PLATFORM } from '@anticrm/platform-ui'
-import presentationPlugin, { AttrModel, ClassModel, ComponentExtension, GroupModel, PresentationService } from '.'
-import { IntlString } from '@anticrm/platform-i18n'
-
-import { deepEqual } from 'fast-equals'
+import { VDoc } from '@anticrm/domains'
 import { Platform } from '@anticrm/platform'
+import core, { CoreService, Unsubscribe, QueryUpdater } from '@anticrm/platform-core'
+import { IntlString } from '@anticrm/platform-i18n'
+import { AnyComponent, CONTEXT_PLATFORM } from '@anticrm/platform-ui'
+import { deepEqual } from 'fast-equals'
+import { getContext, onDestroy } from 'svelte'
+import presentationPlugin, { AttrModel, ClassModel, ComponentExtension, GroupModel, PresentationService } from '.'
 
 export async function getCoreService (): Promise<CoreService> {
   const platform = getContext<Platform>(CONTEXT_PLATFORM)
@@ -42,8 +41,6 @@ export function getUserId (): string {
   const platform = getContext<Platform>(CONTEXT_PLATFORM)
   return platform.getMetadata(core.metadata.WhoAmI) as StringProperty
 }
-
-export type QueryUpdater<T extends Doc> = (_class: Ref<Class<T>>, query: DocumentQuery<T>, options?: FindOptions<T>) => void
 
 /**
  * Perform subscribe to query with some helper finalizer to use
