+++ plugins/workbench/src/components/internal/Application.svelte
@@ -16,7 +16,7 @@
   import type { ItemCreator, WorkbenchApplication } from '../..'
   import workbench from '../..'
 
-  import type { QueryUpdater } from '@anticrm/presentation'
+  import type { QueryUpdater } from '@anticrm/platform-core'
   import ScrollView from '@anticrm/sparkling-controls/src/ScrollView.svelte'
   import CreateForm from './CreateForm.svelte'
   import IconEditBox from '@anticrm/platform-ui/src/components/IconEditBox.svelte'
