+++ plugins/workbench/src/components/internal/Workbench.svelte
@@ -19,7 +19,7 @@
   import Spotlight from './Spotlight.svelte'
   import type { AnyComponent } from '@anticrm/platform-ui'
   import { getUIService, newRouter } from '@anticrm/platform-ui'
-  import type { QueryUpdater } from '@anticrm/presentation'
+  import { QueryUpdater } from '@anticrm/platform-core'
 
   const uiService = getUIService()
